requirements.txt
python-telegram-bot[job-queue]
deepl
pysrt
av
psycopg2-binary
soundfile
aiohttp
requests
websockets
replicate==0.13.0
boto3
aiohttp
aiosmtplib
jinja2
brotli
pycryptodomex
mutagen
yt_dlp